Move python tests to own file
